refactor: Extract internal scalar op support functions.
This will avoid a monolithic file as we add collections.
