+++ mypy.ini
@@ -63,6 +63,9 @@ disallow_any_expr           = False
 [mypy-momento.internal.common._data_client_ops]
 disallow_any_expr           = False
 
+[mypy-momento.internal.common._data_client_scalar_ops]
+disallow_any_expr           = False
+
 [mypy-momento.internal.aio._add_header_client_interceptor]
 disallow_any_expr           = False
