+++ src/momento/internal/synchronous/_scs_data_client.py
@@ -17,14 +17,16 @@ from momento.auth import CredentialProvider
 from momento.config import Configuration
 from momento.internal._utilities import _validate_ttl
 from momento.internal.common._data_client_ops import (
+    get_default_client_deadline,
+    wrap_with_error_handling,
+)
+from momento.internal.common._data_client_scalar_ops import (
     construct_delete_response,
     construct_get_response,
     construct_set_response,
-    get_default_client_deadline,
     prepare_delete_request,
     prepare_get_request,
     prepare_set_request,
-    wrap_with_error_handling,
 )
 from momento.internal.synchronous._scs_grpc_manager import _DataGrpcManager
 from momento.internal.synchronous._utilities import make_metadata
